separate classes from City into separate files
