+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/Model/City.java
@@ -37,26 +37,4 @@ public class City implements Serializable {
     public String getDescription() {
         return this.weather.get(0).description;
     }
-
-    public class Sys implements Serializable {
-        public String country;
-    }
-
-    public class Main implements Serializable {
-        public double temp;
-        public double pressure;
-    }
-
-    public class Wind implements Serializable {
-        public double speed;
-    }
-
-    public class Clouds implements Serializable {
-        public int all;
-    }
-
-    public class Weather implements Serializable {
-        public String description;
-        public String icon;
-    }
 }
