+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/Model/Sys.java
@@ -1,0 +1,7 @@
+package com.rperazzo.weatherapp.Model;
+
+import java.io.Serializable;
+
+public class Sys implements Serializable {
+    public String country;
+}
