+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/Model/Weather.java
@@ -1,0 +1,8 @@
+package com.rperazzo.weatherapp.Model;
+
+import java.io.Serializable;
+
+public class Weather implements Serializable {
+    public String description;
+    public String icon;
+}
