+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/Model/Wind.java
@@ -1,0 +1,7 @@
+package com.rperazzo.weatherapp.Model;
+
+import java.io.Serializable;
+
+public class Wind implements Serializable {
+    public double speed;
+}
