+++ weather-app/app/src/main/java/com/rperazzo/weatherapp/WeatherManager.java
@@ -45,8 +45,6 @@ public class WeatherManager {
             this.list = list;
         }
     }
-
-
 }
