+++ weather-app/app/src/test/java/com/rperazzo/weatherapp/Model/CityTest.java
@@ -2,6 +2,7 @@ package com.rperazzo.weatherapp.Model;
 
 import org.junit.After;
 import org.junit.Before;
+import org.junit.Test;
 
 import static org.junit.Assert.*;
 
@@ -9,9 +10,20 @@ public class CityTest {
 
     @Before
     public void setUp() throws Exception {
+
     }
 
     @After
     public void tearDown() throws Exception {
     }
+
+
+//    @Test
+//    function void testCityAttibutes(){
+//        City c = new City();
+//
+//        c.
+//
+//    }
+
 }
